sysmobts: Re-load the FPGA image and let some time pass before continuing
We are not cleanly shutting down the DSP on exit so let us reset both
the DSP and the FPGA. Also let some time pass to allow the DSP to properly
initialize itself before starting to use it.

This is not a fix but a band-aid for the unclean shutdown.