sysmobts: Read multiple primitives at once but only up to 3
In most cases there are multiple messages ready to be read from
the queue and it is more efficient to read them in one go instead
of going through the select again. On the other hand we need to
avoid stalling the other file descriptors so we are limiting it.